+++ odoo/local-src/mtsmte_sale/models/project.py
@@ -12,3 +12,24 @@ class ProjectProject(models.Model):
     analyze_sample = fields.Text(
         string='Samples To Analyze',
     )
+    resp_one_id = fields.Many2one(
+        'res.users',
+        string='Responsible 1',
+    )
+    resp_two_id = fields.Many2one(
+        'res.users',
+        string='Responsible 2',
+    )
+    send_sample_back = fields.Boolean(
+        string='Send Samples back',
+    )
+    conclusion = fields.Html(
+    )
+
+    def write(self, vals):
+        for proj in self:
+            if 'resp_one_id' in vals:
+                self.message_subscribe_users(vals['resp_one_id'])
+            if 'resp_two_id' in vals:
+                self.message_subscribe_users(vals['resp_two_id'])
+        return super(ProjectProject, self).write(vals)